* Eradicate the use of size_t in favor of a defined type such as uint32_t
git-svn-id: http://opensource.mlba-team.de/svn/xdispatch/trunk@193 9e4c620c-c391-4b63-a7fb-4924991ea895

+++ core/src/apply.c
@@ -32,7 +32,7 @@ struct dispatch_apply_s {
 	_dispatch_apply_function_t	da_func;
 	void	*da_ctxt;
 	size_t	da_iterations;
-	size_t	da_index;
+	uint32_t	da_index;
 	uint32_t	da_thr_cnt;
 	dispatch_semaphore_t da_sema;
 	long	_da_pad1[DISPATCH_CACHELINE_SIZE / sizeof(long)];
